SterlWorld Authentication Login/Register Form
I've finally incorporated a system for visitors to Register/Login and sign out. They sign up via email. 

The next big update will be a User Page. A portal for the visitors to fill in information about themselves for others to view the future. A user dashboard page if you will.